Finished assignment.
Every time a comic is chosen, its id number is immediately saved to a file. When the app starts, it first checks for a file; if there isn't one, the app starts as normal. However, if a file is present, it uses the id to show the comic from the id.